Prepare fake toolbox items to be done for real
