+++ src/toolbox/component_samples_category.ts
@@ -66,232 +66,8 @@ export const category =
                 },
             ],
         },
-        {
-            kind: 'category',
-            name: 'REV Servo',
-            contents: [
-                // def set_position(self, pos: float) -> None:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'None',
-                        args: [
-                            {
-                                name: 'pos',
-                                type: 'float',
-                            },
-                        ],
-                        tooltip: 'Set the servo to a position between 0 and 1.',
-                        importModule: '',
-                        componentClassName: 'rev.Servo',
-                        componentName: 'clawServo',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'clawServo',
-                        FUNC: 'set_position',
-                    },
-                    inputs: {
-                        ARG0: {
-                            block: {
-                                type: 'math_number',
-                                fields: {
-                                    NUM: 0.5,
-                                },
-                            },
-                        },
-                    },
-                },
-                // def set_angle_degrees(self, angle: float) -> None:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'None',
-                        args: [
-                            {
-                                name: 'angle',
-                                type: 'float',
-                            },
-                        ],
-                        tooltip: 'Set the servo to an angle between 0 and 270.',
-                        importModule: '',
-                        componentClassName: 'rev.Servo',
-                        componentName: 'clawServo',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'clawServo',
-                        FUNC: 'set_angle_degrees',
-                    },
-                    inputs: {
-                        ARG0: {
-                            block: {
-                                type: 'math_number',
-                                fields: {
-                                    NUM: 180,
-                                },
-                            },
-                        },
-                    },
-                },
-            ],
-        },
-        {
-            kind: 'category',
-            name: 'REV Smart Motor',
-            contents: [
-                // def set_speed(self, speed: float) -> None:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'None',
-                        args: [
-                            {
-                                name: 'speed',
-                                type: 'float',
-                            },
-                        ],
-                        tooltip: 'Set the motor to a speed between -1 and 1.',
-                        importModule: '',
-                        componentClassName: 'rev.SmartMotor',
-                        componentName: 'leftMotor',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'leftMotor',
-                        FUNC: 'set_speed',
-                    },
-                    inputs: {
-                        ARG0: {
-                            block: {
-                                type: 'math_number',
-                                fields: {
-                                    NUM: 0.8,
-                                },
-                            },
-                        },
-                    },
-                },
-                // def set_angle_degrees(self, angle: float) -> None:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'None',
-                        args: [
-                            {
-                                name: 'angle',
-                                type: 'float',
-                            },
-                        ],
-                        tooltip: 'Set the motor to an angle between 0 and 360.',
-                        importModule: '',
-                        componentClassName: 'rev.SmartMotor',
-                        componentName: 'leftMotor',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'leftMotor',
-                        FUNC: 'set_angle_degrees',
-                    },
-                    inputs: {
-                        ARG0: {
-                            block: {
-                                type: 'math_number',
-                                fields: {
-                                    NUM: 180,
-                                },
-                            },
-                        },
-                    },
-                },
-                // def get_num_relative_encoder_ticks(self) -> int:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'int',
-                        args: [],
-                        tooltip: 'Get the number of relative motor ticks since reset of encoder.',
-                        importModule: '',
-                        componentClassName: 'rev.SmartMotor',
-                        componentName: 'leftMotor',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'leftMotor',
-                        FUNC: 'get_num_relative_encoder_ticks',
-                    },
-                    inputs: {},
-                },
-                // def get_angle_degrees(self) -> float:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'float',
-                        args: [],
-                        tooltip: 'Get the angle position of the motor.',
-                        importModule: '',
-                        componentClassName: 'rev.SmartMotor',
-                        componentName: 'leftMotor',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'leftMotor',
-                        FUNC: 'get_angle_degrees',
-                    },
-                    inputs: {},
-                },
-                // def reset_relative_encoder(self) -> None:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'None',
-                        args: [],
-                        tooltip: 'Reset the relative encoder value to 0.',
-                        importModule: '',
-                        componentClassName: 'rev.SmartMotor',
-                        componentName: 'leftMotor',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'leftMotor',
-                        FUNC: 'reset_relative_encoder',
-                    },
-                    inputs: {},
-                },
-            ]
-        },
-        {
-            kind: 'category',
-            name: 'REV Touch Sensor',
-            contents: [ 
-                // def is_pressed(self) -> bool:
-                {
-                    kind: 'block',
-                    type: 'mrc_call_python_function',
-                    extraState: {
-                        functionKind: 'instance_component',
-                        returnType: 'bool',
-                        args: [],
-                        tooltip: 'Return whether the touch sensor is pressed.',
-                        importModule: '',
-                        componentClassName: 'rev.TouchSensor',
-                        componentName: 'frontTouch',
-                    },
-                    fields: {
-                        COMPONENT_NAME: 'frontTouch',
-                        FUNC: 'is_pressed',
-                    },
-                    inputs: {},
-                },
-            ],
-        },
+        
+  
         {
             kind: 'category',
             name: 'SparkFun LED Stick',
